Reformat text from the WICG draft and as kramdown-rfc compatible markdown.
The text is mostly preserved verbatim, though since auto-linking in kramdown-rfc is much more limited, not all references (especially to WHATWG Infra) are linked. The term "list" is replaced with "array" (as in RFC 8941) only in contexts where it is applied directly to the result of RFC 8941, which uses the term "list" to refer to a specific kind of structured field value which is not meant here (and it may be confusing which is meant, even with the remarks about WHATWG notation and conventions).

Some formatting is also lost due to the constraints of the Markdown -> HTML -> RFC XML -> {HTML, TXT} pipeline used for modern Internet-Draft processing.

No additional text is yet added, e.g., related to integration with RFC 9111 (HTTP-CACHING) or IANA registration.